separated telegram function
